app_requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.10.0
Pillow>=10.0.0
numpy>=1.24.0
